refactor: cargo_generate_lockfile -> cargo_update.
