+++ src/cargo/ops/mod.rs
@@ -7,10 +7,6 @@ pub use self::cargo_compile::{
 pub use self::cargo_compile::{CompileFilter, FilterRule, LibRule, Packages};
 pub use self::cargo_doc::{doc, DocOptions, OutputFormat};
 pub use self::cargo_fetch::{fetch, FetchOptions};
-pub use self::cargo_generate_lockfile::generate_lockfile;
-pub use self::cargo_generate_lockfile::print_lockfile_changes;
-pub use self::cargo_generate_lockfile::update_lockfile;
-pub use self::cargo_generate_lockfile::UpdateOptions;
 pub use self::cargo_install::{install, install_list};
 pub use self::cargo_new::{init, new, NewOptions, NewProjectKind, VersionControl};
 pub use self::cargo_output_metadata::{output_metadata, ExportInfo, OutputMetadataOptions};
@@ -20,6 +16,10 @@ pub use self::cargo_read_manifest::{read_package, read_packages};
 pub use self::cargo_run::run;
 pub use self::cargo_test::{run_benches, run_tests, TestOptions};
 pub use self::cargo_uninstall::uninstall;
+pub use self::cargo_update::generate_lockfile;
+pub use self::cargo_update::print_lockfile_changes;
+pub use self::cargo_update::update_lockfile;
+pub use self::cargo_update::UpdateOptions;
 pub use self::fix::{fix, fix_exec_rustc, fix_get_proxy_lock_addr, FixOptions};
 pub use self::lockfile::{load_pkg_lockfile, resolve_to_string, write_pkg_lockfile};
 pub use self::registry::modify_owners;
@@ -44,7 +44,6 @@ pub(crate) mod cargo_compile;
 pub mod cargo_config;
 mod cargo_doc;
 mod cargo_fetch;
-mod cargo_generate_lockfile;
 mod cargo_install;
 mod cargo_new;
 mod cargo_output_metadata;
@@ -55,6 +54,7 @@ pub mod cargo_remove;
 mod cargo_run;
 mod cargo_test;
 mod cargo_uninstall;
+mod cargo_update;
 mod common_for_install_and_uninstall;
 mod fix;
 pub(crate) mod lockfile;
